- Tudo Feito + css

+++ siteiscte/votacao/views.py
@@ -70,7 +70,7 @@ def voto(request, questao_id):
     else:
         if not request.user.is_superuser:
             num_grupo = request.user.aluno.grupo[-2:]
-            if request.user.aluno.votos < int(num_grupo):
+            if request.user.aluno.votos < int(num_grupo) + 10:
                 opcao_seleccionada.votos += 1
                 opcao_seleccionada.save()
                 request.user.aluno.votos += 1
@@ -88,7 +88,6 @@ def createquestion(request):
     try:
         questaotexto = request.POST['qname']
     except(KeyError, Questao.DoesNotExist):
-        # Se não demos post de uma questão então mostra a página para criar a questão
         return render(request, 'votacao/createquestion.html')
     else:
         questao = Questao(questao_texto=questaotexto, pub_data=datetime.datetime.now())
@@ -97,17 +96,16 @@ def createquestion(request):
 
 
 def deletequestion(request, questao_id):
-    questao = get_object_or_404(Questao, pk=questao_id)
-    questao.delete()
+    if request.user.is_superuser and request.user.is_authenticated:
+        questao = get_object_or_404(Questao, pk=questao_id)
+        questao.delete()
     return HttpResponseRedirect(reverse('votacao:index'))
-
 
 def createoption(request, questao_id):
     questao = get_object_or_404(Questao, pk=questao_id)
     try:
         opcaotexto = request.POST['opname']
     except(KeyError, Opcao.DoesNotExist):
-        # Se não demos post de uma opção então mostra a página para criar a opção
         return render(request, 'votacao/createoption.html', {'questao': questao})
     else:
         opcao = Opcao(questao=questao, opcao_texto=opcaotexto, votos=0)
@@ -120,7 +118,6 @@ def deleteoption(request, questao_id):
     try:
         opcao_seleccionada = questao.opcao_set.get(pk=request.POST['opcao'])
     except (KeyError, Opcao.DoesNotExist):
-        # Apresenta de novo o form
         return render(request, 'votacao/detalhe.html', {
             'questao': questao, 'error_message': "Não escolheu uma opção", })
     else:
